Bug 536324 part 2 - Change implementers of nsIChannel. r=sworkman
UltraBlame original commit: 90d8a63e061cc05f84d5f4c87dd9b2ebded439b9

+++ modules/libjar/nsJARChannel.h
@@ -71,7 +71,7 @@ private:
     
 
     uint32_t                        mContentDisposition;
-    int32_t                         mContentLength;
+    int64_t                         mContentLength;
     uint32_t                        mLoadFlags;
     nsresult                        mStatus;
     bool                            mIsPending;
